Adding first version of Runner and Model folder
