+++ backend/orders/src/bootstrap.kt
@@ -1,5 +1,6 @@
 package challenge
 
+import challenge.model.Product
 import java.lang.Exception
 import java.util.*
 
@@ -57,8 +58,6 @@ data class Invoice(val order: Order) {
     val shippingAddress: Address = order.address
 }
 
-data class Product(val name: String, val type: ProductType, val price: Double)
-
 enum class ProductType {
     PHYSICAL,
     BOOK,
@@ -70,19 +69,19 @@ class Address
 
 class Customer
 
-fun main(args : Array<String>) {
-    val shirt = Product("Flowered t-shirt", ProductType.PHYSICAL, 35.00)
-    val netflix = Product("Familiar plan", ProductType.MEMBERSHIP, 29.90)
-    val book = Product("The Hitchhiker's Guide to the Galaxy", ProductType.BOOK, 120.00)
-    val music = Product("Stairway to Heaven", ProductType.DIGITAL, 5.00)
-
-    val order = Order(Customer(), Address())
-
-    order.addProduct(shirt, 2)
-    order.addProduct(netflix, 1)
-    order.addProduct(book, 1)
-    order.addProduct(music, 1)
-
-    order.pay(CreditCard("43567890-987654367"))
-    // now, how to deal with shipping rules then?
-}
+//fun main(args : Array<String>) {
+//    val shirt = Product("Flowered t-shirt", ProductType.PHYSICAL, 35.00)
+//    val netflix = Product("Familiar plan", ProductType.MEMBERSHIP, 29.90)
+//    val book = Product("The Hitchhiker's Guide to the Galaxy", ProductType.BOOK, 120.00)
+//    val music = Product("Stairway to Heaven", ProductType.DIGITAL, 5.00)
+//
+//    val order = Order(Customer(), Address())
+//
+//    order.addProduct(shirt, 2)
+//    order.addProduct(netflix, 1)
+//    order.addProduct(book, 1)
+//    order.addProduct(music, 1)
+//
+//    order.pay(CreditCard("43567890-987654367"))
+//    // now, how to deal with shipping rules then?
+//}
