+++ backend/orders/src/model/Product.kt
@@ -1,0 +1,6 @@
+package challenge.model
+
+import challenge.ProductType
+
+
+data class Product(val name: String, val type: ProductType, val price: Double)
